Employee route and User tests
Added two tests for the employee route (guest user is not supposed to
access the route, and an existing user must be able to access it).
Added a test for the login route (users with incorrect password cannot
authenticate).
Fixed use of html reserved characters in login page (replaced with
equivalent html entity).