UserType Factory and other minor fixes
The addition of an "acronym" field was breaking the UserType factory,
added. Switch other generators to the pt_BR version.
EmployeeFactory was breaking because of the limit of characters
in the postal code field, fixed by removing the dash on the CEP code.
GenderFactory was breaking because of the limit of characters in the
field name field, fixed by using the Str::limit helper.