Employee Seeder (WIP)
Added an Employee Seeder, still work in progress because the
requirements are unclear. Currently it add a "Coordenador" for each
course, with Bond, and Pole set to "SEAD", since coordinators are shared
accross all pole/regions.

This Seeder can later be merged into the EmployeeSeeder or replace it
(once it's the requirements are clearer and the Seeder is completed).