Employee Seeder "Tutor Presencial"
According to the client accounts, each Course has a Tutor (Tutor
Presencial) in each Pole. Creating them for each Pole and Course on the
Seeders would amount to 700+ tutors.
To avoid too much clutter in the example data, the mumber of Poles and
Courses was limited to only three of each (but this can be easily
altered in the Seeder file, before the final version).