Tests for CRUD operations in Pole
Checks that guest users cannot reach any of the internal routes for the
Pole related areas of the application. Also checks they get redirected
to the login page.

Checks that authenticated users can perform all CRUD operations related
to the Pole areas of the application.

Minor change to the name of class TextHelperTest, to conform to psr-12
and prevent composer from showing the annoying warning in every update.